Download Datavyu File from Databrary

Download Datavyu File from Databrary

download_opf(
  session_id = 9807,
  asset_id = 117035,
  file_name = NULL,
  vb = FALSE,
  rq = NULL
)

Arguments

session_id

An integer. The unique identifier for a Databrary session that contains a Datavyu file. Default is 9807.

asset_id.

An integer. The unique identifier for the Datavyu file in the Databrary session. Default is 117035.

file_name.

A string. The name of the file to be downloaded. Default is NULL. If the file name is NULL a default one is generated.

vb.

A logical value. Provide verbose information at the R console. Default is FALSE.

rq.

An httr2 request object. Default is NULL. This generates a new default request using databraryr::make_default_request(). If you are accessing Datavyu files from a volume or session accessible only to authorized users, you should authenticate to Databrary and use the httr2 request associated with that authentication.

Value

A string with the full file name of the downloaded file.

Examples

# \donttest{
download_opf() # Downloads a test Datavyu file from Databrary volume 1.
#> [1] "/var/folders/n3/6wm9hw792y344j61fqh2rw94mcvkj6/T//Rtmp8mHv5m/9807-117035-2024-02-21-0858-55.opf"
# }