Convert Extracted Datavyu File To CSV.

Convert Extracted Datavyu File To CSV.

to_csv(
  dv_dir = ".",
  dv_fn = "db",
  out_fn = paste0(dv_dir, "/tmp.csv"),
  auto_write_over = FALSE,
  code_regex = "^([a-zA-Z_]+[0-9]*[a-zA-Z_]*[0-9]*)",
  code_type_regex = "([a-zA-Z]+)$",
  time_regex = "([0-9]{2}:[0-9]{2}:[0-9]{2}:[0-9]{3})",
  code_values_regex = "\\(([a-zA-Z ?,.'/0-9;!|~`]+)\\)$",
  convert_times = TRUE,
  vb = FALSE
)

Arguments

dv_dir

A character string. The directory to extract the Datavyu file. Default value is the current directory ".".

dv_fn

A character string. The name of the Datavyu code file. Defaults to 'db'.

out_fn

A character string. The name of the output CSV file. Default 'tmp.csv' in the directory specified by dv_dir.

auto_write_over

A logical value. If TRUE, new output file overwrites the old. Default is FALSE.

code_regex

A character string. A regular expression to extract codes from a Datavyu file.

code_type_regex

A character string. A regular expression to extract code types from a Datavyu file.

time_regex

A character string. A regular expression to extract onset/offset times from a Datavyu file.

code_values_regex

A character string. A regular expression to extract code values from Datavyu file.

convert_times

A logical value. If TRUE (default), converts the onset and offset fields to lubridate-compatible dates and times. Default is TRUE.

vb

A boolean value. If TRUE, provides verbose output. Default is FALSE.

Examples

# \donttest{
to_csv()
#> Warning: cannot open file './db': No such file or directory
#> Error in file(paste0(dv_dir, "/", dv_fn), "r"): cannot open the connection
# }